feat: 쉼표만 지운 경우, 쉼표 앞으로 커서 이동하는 기능 추가
